requirements.txt
aiogram
requests
PyJWT
aiohttp
python-dotenv
jwt
openai
tiktoken
opencv-python
sqlalchemy